refactor: move update_dashboard_with_gst_logs to utils
(cherry picked from commit ffbde4bab0349a30f15bda3e96d32dbe614f2a29)
