+++ india_compliance/gst_india/overrides/purchase_receipt.py
@@ -3,15 +3,12 @@ import frappe
 from india_compliance.gst_india.overrides.purchase_invoice import (
     set_ineligibility_reason,
 )
-from india_compliance.gst_india.overrides.sales_invoice import (
-    update_dashboard_with_gst_logs,
-)
 from india_compliance.gst_india.overrides.transaction import (
     ignore_gst_validations,
     validate_mandatory_fields,
     validate_transaction,
 )
-from india_compliance.gst_india.utils import is_api_enabled
+from india_compliance.gst_india.utils import is_api_enabled, update_dashboard_with_gst_logs
 from india_compliance.gst_india.utils.e_waybill import get_e_waybill_info
